--- app/src/main/java/com/example/andr_finalproject_modul_19_part_1/MainActivity.kt
@@ -1,13 +1,0 @@
-package com.example.andr_finalproject_modul_19_part_1
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
-        }
-    }
