+++ app/src/main/res/drawable/corner_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:startColor="@android:color/holo_orange_light"
+        android:endColor="@color/colorPrimaryDark"
+        android:angle="45"/>
+    <corners android:radius="15dp"/>
+
+</shape>
