+++ app/src/main/res/layout-sw600dp/activity_main.xml
@@ -1,0 +1,174 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:gravity="center_vertical">
+
+    <!--Контейнер под нижнее меню-->
+
+    <FrameLayout
+        android:id="@+id/downMenu"
+        android:layout_width="match_parent"
+        android:layout_height="123dp"
+        android:background="@color/colorGreen"
+        app:layout_constraintBottom_toBottomOf="parent" />  <!--Привязка к низу экрана-->
+
+    <!--Контейнер под верхнее меню и кнопка "Поиск фильмов"-->
+
+    <LinearLayout
+        android:id="@+id/topMenu"
+        android:layout_width="match_parent"
+        android:layout_height="110dp"
+        android:text="@string/text_on_topMenu"
+        android:textSize="40sp"
+        android:background="@color/colorBlue"
+        android:orientation="vertical"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent">
+
+           <androidx.appcompat.widget.AppCompatButton
+               android:id="@+id/button0"
+               android:layout_width="match_parent"
+               android:layout_height="40dp"
+               android:layout_marginTop="50dp"
+               android:background="@color/colorSky"
+               android:insetLeft="0dp"
+               android:text="@string/text_on_topMenu"
+               android:textColor="@color/white"
+               android:textSize="18sp"
+               android:gravity="center_vertical|start"
+               android:paddingStart="10dp"/> <!--Отступ слева для текста-->
+
+    </LinearLayout>
+
+    <!--Контейнер под меню с кнопками резиновой высотой(0 dp)-->
+
+    <LinearLayout
+        android:id="@+id/buttonMenu"
+        android:orientation="vertical"
+        android:layout_width="600dp"
+        android:layout_height="wrap_content"
+
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/topMenu"
+        app:layout_constraintBottom_toTopOf="@+id/downMenu"
+        android:background="@drawable/corner_background" >
+
+    <Button
+        android:id="@+id/button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/colorYellow"
+        android:background="?selectableItemBackgroundBorderless"
+        android:onClick="onClickToast"
+        android:text="@string/btn_menu_1"/>
+
+    <Button
+        android:id="@+id/button2"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/colorYellow"
+        android:background="?selectableItemBackgroundBorderless"
+        android:onClick="onClickToast2"
+        android:text="@string/btn_menu_2" />
+
+    <Button
+        android:id="@+id/button3"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/colorYellow"
+        android:background="?selectableItemBackgroundBorderless"
+        android:onClick="onClickToast3"
+        android:text="@string/btn_menu_3" />
+
+    <Button
+        android:id="@+id/button4"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/colorYellow"
+        android:background="?selectableItemBackgroundBorderless"
+        android:onClick="onClickToast4"
+        android:text="@string/btn_menu_4" />
+
+    <Button
+        android:id="@+id/button5"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textColor="@color/colorYellow"
+        android:background="?selectableItemBackgroundBorderless"
+        android:onClick="onClickToast5"
+        android:text="@string/btn_menu_5" />
+
+    </LinearLayout>
+
+    <!--Контейнер под меню с картинками-->
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/top_bar"
+        android:layout_width="match_parent"
+        android:layout_height="250dp"
+        android:orientation="horizontal"
+        android:layout_margin="5dp"
+        app:cardCornerRadius="15dp"
+        android:elevation="5dp"
+
+        app:layout_constraintTop_toBottomOf="@+id/topMenu">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="horizontal">
+
+        <ImageView
+            android:id="@+id/imageView"
+            android:scaleType="centerCrop"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            app:srcCompat="@drawable/gladiator" />
+
+        <ImageView
+            android:id="@+id/imageView2"
+            android:scaleType="centerCrop"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            app:srcCompat="@drawable/interstellar" />
+
+        <ImageView
+            android:id="@+id/imageView3"
+            android:scaleType="centerCrop"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1 "
+            app:srcCompat="@drawable/backtofuture" />
+
+        <ImageView
+            android:id="@+id/imageView4"
+            android:scaleType="centerCrop"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            app:srcCompat="@drawable/darkknightbuilding" />
+        </LinearLayout>
+
+    </androidx.cardview.widget.CardView>
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="5dp"
+
+        android:text="@string/disclaimer"
+        android:autoLink="web"
+        android:linksClickable="true"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/top_bar" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
