+++ app/src/main/res/layout/activity_main.xml
@@ -3,35 +3,173 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center_vertical">
+    android:gravity="center_vertical"
+    android:orientation="vertical">
 
     <!--Контейнер под нижнее меню-->
+
     <FrameLayout
         android:id="@+id/downMenu"
         android:layout_width="match_parent"
-        android:layout_height="70dp"
+        android:layout_height="123dp"
         android:background="@color/colorGreen"
-        app:layout_constraintBottom_toBottomOf="parent" /> <!--Привязка к низу экрана-->
+        app:layout_constraintBottom_toBottomOf="parent" />
 
-    <!--Контейнер под верхнее меню-->
-    <FrameLayout
+    <!--Контейнер под верхнее меню и кнопка "Поиск фильмов"-->
+
+    <LinearLayout
         android:id="@+id/topMenu"
         android:layout_width="match_parent"
-        android:layout_height="70dp"
+        android:layout_height="110dp"
         android:background="@color/colorBlue"
-        app:layout_constraintTop_toTopOf="parent"/> <!--Привязка к верху экрана-->
+        android:orientation="vertical"
+        android:text="@string/text_on_topMenu"
+        android:textSize="40sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
 
-    <!--Контейнер под меню с кнопками резиновой высотой(0 dp)-->
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/button0"
+            android:layout_width="match_parent"
+            android:layout_height="40dp"
+            android:layout_marginTop="50dp"
+            android:background="@color/colorSky"
+            android:gravity="center_vertical|start"
+            android:insetLeft="0dp"
+            android:paddingStart="10dp"
+            android:text="@string/text_on_topMenu"
+            android:textColor="@color/white"
+            android:textSize="18sp" /> <!--Отступ слева для текста-->
+
+    </LinearLayout>
+
+    <!--Контейнер под меню с кнопками резиновой высотой(wrap_content dp)-->
+
     <LinearLayout
         android:id="@+id/buttonMenu"
-        android:orientation="vertical"
         android:layout_width="300dp"
-        android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/topMenu"
+        android:layout_height="wrap_content"
+        android:background="@drawable/corner_background"
+
+        android:orientation="vertical"
         app:layout_constraintBottom_toTopOf="@+id/downMenu"
-        android:background="@color/colorYellow" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/topMenu">
+
+        <Button
+            android:id="@+id/button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:onClick="onClickToast"
+            android:text="@string/btn_menu_1"
+            android:textColor="@color/colorYellow" />
+
+        <Button
+            android:id="@+id/button2"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:onClick="onClickToast2"
+            android:text="@string/btn_menu_2"
+            android:textColor="@color/colorYellow" />
+
+        <Button
+            android:id="@+id/button3"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:onClick="onClickToast3"
+            android:text="@string/btn_menu_3"
+            android:textColor="@color/colorYellow" />
+
+        <Button
+            android:id="@+id/button4"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:onClick="onClickToast4"
+            android:text="@string/btn_menu_4"
+            android:textColor="@color/colorYellow" />
+
+        <Button
+            android:id="@+id/button5"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:onClick="onClickToast5"
+            android:text="@string/btn_menu_5"
+            android:textColor="@color/colorYellow" />
+
+    </LinearLayout>
+
+    <!--Контейнер под меню с картинками-->
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/top_bar"
+        android:layout_width="match_parent"
+        android:layout_height="100dp"
+        android:layout_margin="5dp"
+        android:elevation="5dp"
+        android:orientation="horizontal"
+        app:cardCornerRadius="15dp"
+        app:layout_constraintTop_toBottomOf="@+id/topMenu">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="horizontal">
+
+            <ImageView
+                android:id="@+id/imageView"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:scaleType="centerCrop"
+                app:srcCompat="@drawable/gladiator" />
+
+            <ImageView
+                android:id="@+id/imageView2"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:scaleType="centerCrop"
+                app:srcCompat="@drawable/interstellar" />
+
+            <ImageView
+                android:id="@+id/imageView3"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1 "
+                android:scaleType="centerCrop"
+                app:srcCompat="@drawable/backtofuture" />
+
+            <ImageView
+                android:id="@+id/imageView4"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:scaleType="centerCrop"
+                app:srcCompat="@drawable/darkknightbuilding" />
+        </LinearLayout>
+
+    </androidx.cardview.widget.CardView>
+
+    <!--Текст под картинками с адресом откуда брали-->
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="5dp"
+
+        android:autoLink="web"
+        android:linksClickable="true"
+        android:text="@string/disclaimer"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/top_bar" />
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
