+++ app/src/main/res/values/themes.xml
@@ -1,9 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.ANDR_finalproject_modul_19_part_1" parent="Theme.Material3.DayNight.NoActionBar">
+    <style name="Base.Theme.AndroidAppFilmProject" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
     </style>
 
-    <style name="Theme.ANDR_finalproject_modul_19_part_1" parent="Base.Theme.ANDR_finalproject_modul_19_part_1" />
+    <style name="Theme.AndroidAppFilmProject" parent="Base.Theme.AndroidAppFilmProject" />
 </resources>
